ref: split models in files
