+++ src/api/blocks.rs
@@ -1,8 +1,8 @@
 use crate::http::client::Client;
 use std::borrow::Borrow;
 
-use crate::api::models_new::block::Block;
-use crate::api::models_new::transaction::Transaction;
+use crate::api::models::block::Block;
+use crate::api::models::transaction::Transaction;
 
 use crate::api::Result;
