+++ src/api/models/asset.rs
