+++ src/api/models/delegate.rs
@@ -1,0 +1,44 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+use crate::api::models::timestamp::Timestamp;
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Delegate {
+    pub username: String,
+    pub address: String,
+    pub public_key: String,
+    pub votes: u64,
+    pub rank: u32,
+    pub blocks: Blocks,
+    pub production: Production,
+    pub forged: DelegateForged,
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+pub struct Blocks {
+    pub produced: u64,
+    pub missed: u64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub last: Option<Last>,
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+pub struct Production {
+    pub approval: f64,
+    pub productivity: f64,
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+pub struct DelegateForged {
+    pub rewards: u64,
+    pub fees: u64,
+    pub total: u64,
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+pub struct Last {
+    pub id: String,
+    pub timestamp: Timestamp,
+}
