+++ src/api/models/shared.rs
@@ -1,0 +1,37 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RequestError {
+    pub status_code: i16,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub message: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Response<T> {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub meta: Option<Meta>,
+    pub data: T,
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Meta {
+    pub total_count_is_estimate: bool,
+    pub count: u32,
+    pub page_count: u32,
+    pub total_count: u32,
+    pub next: Option<String>,
+    pub previous: Option<String>,
+    #[serde(rename = "self")]
+    pub self_url: String,
+    pub first: String,
+    pub last: Option<String>,
+}
