+++ src/api/models/transaction.rs
@@ -1,8 +1,8 @@
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::api::models_new::asset::Asset;
-use crate::api::models_new::timestamp::Timestamp;
+use crate::api::models::asset::Asset;
+use crate::api::models::timestamp::Timestamp;
 
 use crate::common::deserialize_u64_as_number_or_string;
