+++ src/error.rs
@@ -1,4 +1,4 @@
-use crate::api::models::RequestError;
+use crate::api::models::shared::RequestError;
 use reqwest;
 use serde_json;
 use std::error;
